requirements.txt
apache-airflow
apache-airflow-providers-google
dbt-core
dbt-bigquery
google-cloud-storage
pandas
requests
SQLAlchemy
tenacity